Refactor KeyboardPressInput component out and add tests for it
